[System] grab "Share usage and diagnostic data"
