+++ ui/component/settingSystem/index.js
@@ -1,7 +1,14 @@
 import { connect } from 'react-redux';
 import { doWalletStatus, selectWalletIsEncrypted } from 'lbry-redux';
-import { doClearCache, doNotifyDecryptWallet, doNotifyEncryptWallet, doNotifyForgetPassword } from 'redux/actions/app';
+import {
+  doClearCache,
+  doNotifyDecryptWallet,
+  doNotifyEncryptWallet,
+  doNotifyForgetPassword,
+  doToggle3PAnalytics,
+} from 'redux/actions/app';
 import { doSetDaemonSetting, doClearDaemonSetting, doFindFFmpeg } from 'redux/actions/settings';
+import { selectAllowAnalytics } from 'redux/selectors/app';
 import { selectDaemonSettings, selectFfmpegStatus, selectFindingFFmpeg } from 'redux/selectors/settings';
 import { selectUserVerifiedEmail } from 'redux/selectors/user';
 import SettingSystem from './view';
@@ -12,6 +19,7 @@ const select = (state) => ({
   findingFFmpeg: selectFindingFFmpeg(state),
   walletEncrypted: selectWalletIsEncrypted(state),
   isAuthenticated: selectUserVerifiedEmail(state),
+  allowAnalytics: selectAllowAnalytics(state),
 });
 
 const perform = (dispatch) => ({
@@ -23,6 +31,7 @@ const perform = (dispatch) => ({
   decryptWallet: () => dispatch(doNotifyDecryptWallet()),
   updateWalletStatus: () => dispatch(doWalletStatus()),
   confirmForgetPassword: (modalProps) => dispatch(doNotifyForgetPassword(modalProps)),
+  toggle3PAnalytics: (allow) => dispatch(doToggle3PAnalytics(allow)),
 });
 
 export default connect(select, perform)(SettingSystem);
